+++ apps/examples/streaming-wallet-extension/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @lightsparkdev/streaming-wallet-extension
 
+## 1.0.35
+
+### Patch Changes
+
+- Updated dependencies [1372f3a]
+- Updated dependencies [24782f5]
+  - @lightsparkdev/lightspark-sdk@1.1.4
+  - @lightsparkdev/core@1.0.7
+
+## 1.0.34
+
+### Patch Changes
+
+- Updated dependencies [baeb7a1]
+- Updated dependencies [ffcedbe]
+- Updated dependencies [d9f6d5b]
+- Updated dependencies [baeb7a1]
+  - @lightsparkdev/lightspark-sdk@1.1.3
+  - @lightsparkdev/core@1.0.6
+
 ## 1.0.33
 
 ### Patch Changes