+++ packages/lightspark-cli/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @lightsparkdev/lightspark-cli
 
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies [1372f3a]
+- Updated dependencies [24782f5]
+  - @lightsparkdev/lightspark-sdk@1.1.4
+  - @lightsparkdev/core@1.0.7
+
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies [baeb7a1]
+- Updated dependencies [ffcedbe]
+- Updated dependencies [d9f6d5b]
+- Updated dependencies [baeb7a1]
+  - @lightsparkdev/lightspark-sdk@1.1.3
+  - @lightsparkdev/core@1.0.6
+
 ## 0.0.23
 
 ### Patch Changes