+++ packages/lightspark-sdk/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @lightsparkdev/lightspark-sdk
 
+## 1.1.4
+
+### Patch Changes
+
+- 1372f3a: Move transaction polling into new client method
+- Updated dependencies [24782f5]
+  - @lightsparkdev/core@1.0.7
+
+## 1.1.3
+
+### Patch Changes
+
+- baeb7a1: [lightspark-sdk] Fix improper fundNode var ref and add tests
+- Updated dependencies [ffcedbe]
+- Updated dependencies [d9f6d5b]
+- Updated dependencies [baeb7a1]
+  - @lightsparkdev/core@1.0.6
+
 ## 1.1.2
 
 ### Patch Changes