+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @lightsparkdev/react-native
 
+## 0.2.22
+
+### Patch Changes
+
+- Updated dependencies [24782f5]
+  - @lightsparkdev/core@1.0.7
+
+## 0.2.21
+
+### Patch Changes
+
+- Updated dependencies [ffcedbe]
+- Updated dependencies [d9f6d5b]
+- Updated dependencies [baeb7a1]
+  - @lightsparkdev/core@1.0.6
+
 ## 0.2.20
 
 ### Patch Changes