+++ packages/react-wallet/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @lightsparkdev/react-wallet
 
+## 0.6.31
+
+### Patch Changes
+
+- Updated dependencies [24782f5]
+  - @lightsparkdev/core@1.0.7
+  - @lightsparkdev/wallet-sdk@0.7.4
+
+## 0.6.30
+
+### Patch Changes
+
+- Updated dependencies [ffcedbe]
+- Updated dependencies [d9f6d5b]
+- Updated dependencies [baeb7a1]
+  - @lightsparkdev/core@1.0.6
+  - @lightsparkdev/wallet-sdk@0.7.3
+
 ## 0.6.29
 
 ### Patch Changes