+++ packages/wallet-cli/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @lightsparkdev/wallet-cli
 
+## 0.0.40
+
+### Patch Changes
+
+- Updated dependencies [24782f5]
+  - @lightsparkdev/core@1.0.7
+  - @lightsparkdev/wallet-sdk@0.7.4
+
+## 0.0.39
+
+### Patch Changes
+
+- Updated dependencies [ffcedbe]
+- Updated dependencies [d9f6d5b]
+- Updated dependencies [baeb7a1]
+  - @lightsparkdev/core@1.0.6
+  - @lightsparkdev/wallet-sdk@0.7.3
+
 ## 0.0.38
 
 ### Patch Changes